chore: export cmd

+++ src/commands/run.ts
@@ -116,7 +116,7 @@ export default class Run extends RunBase {
 
     let child = null
     process.on('SIGINT', () => {
-      child.kill('SIGINT')
+      child && child.kill('SIGINT')
       process.exit(0)
     })
 
@@ -127,7 +127,7 @@ export default class Run extends RunBase {
       })
 
       rl.on('SIGINT', () => {
-        child.kill('SIGINT')
+        child && child.kill('SIGINT')
         process.exit(0)
       })
     }
